improve comment getfixturevalue, move test_scoped_fixture_teardown_order to testing/python/fixtures.py
